Soss.jl

Soss.MarkovChain

Soss.PySymPyModule

Soss.Do

Soss.advancedHMC

Soss.after

Soss.before

Soss.dynamicHMC

Soss.predictive

Soss.prior

Soss.prune

Soss.MarkovChain — Type

MarkovChain

MarkovChain(pars, step) defines a Markov Chain with global parameters pars and transition kernel step. Here, pars is a named tuple, and step is a Soss model that takes arguments (pars, state) and returns a next value containing the new pars and state.

NOTE: This is experimental, and may change in the near future.

mstep = @model pars,state begin
    σ = pars.σ
    x0 = state.x
    x ~ Normal(x0, σ)
    next = (pars=pars, state=(x=x,))
end;

m = @model s0 begin
    σ ~ Exponential()
    pars = (σ=σ,)
    chain ~ MarkovChain(pars, mstep(pars=pars, state=s0))
end;

r = rand(m(s0=(x=2,),));

for s in Iterators.take(r.chain,3)
    println(s)
end

# output

(x = -6.596883394256064,)
(x = 0.48200039561318864,)
(x = -2.838556784903994,)

Soss.Do — Method

Do(m, xs...)

Returns a model transformed by adding xs... to arguments. The remainder of the body remains the same, consistent with Judea Pearl's "Do" operator. Unneeded arguments are trimmed.

Examples

m = @model (n, k) begin
    β ~ Gamma()
    α ~ Gamma()
    θ ~ Beta(α, β)
    x ~ Binomial(n, θ)
    z ~ Binomial(k, α / (α + β))
end;
Do(m, :θ)

# output
@model (n, k, θ) begin
        β ~ Gamma()
        α ~ Gamma()
        x ~ Binomial(n, θ)
        z ~ Binomial(k, α / (α + β))
    end

Soss.advancedHMC — Method

advancedHMC(m, data, N = 1000; n_adapts = 1000)

Draw N samples from the posterior distribution of parameters defined in Soss model m, conditional on data. Samples are drawn using Hamiltonial Monte Carlo (HMC) from the advancedHMC.jl package.

Keywords

n_adapts = 1000: The number of interations used to set HMC parameters.

Returns a tuple of length 2:

Samples from the posterior distribution of parameters.

Sample summary statistics.

Example

using Random
Random.seed!(42);

m = @model x begin
    β ~ Normal()
    yhat = β .* x
    y ~ For(eachindex(x)) do j
        Normal(yhat[j], 2.0)
    end
end

x = randn(3);
truth = rand(m(x=x));

post = advancedHMC(m(x=x), (y=truth.y,));
E_β = mean(post[1])[1]

println("true β: " * string(round(truth.β, digits=2)))
println("Posterior mean β: " * string(round(E_β, digits=2)))

# output
true β: -0.3
Posterior mean β: -0.25

Soss.dynamicHMC — Function

dynamicHMC(
    rng::AbstractRNG,
    m::JointDistribution,
    _data,
    N::Int = 1000;
    method = logpdf,
    ad_backend = Val(:ForwardDiff),
    reporter = DynamicHMC.NoProgressReport(),
    kwargs...)

Draw N samples from the posterior distribution of parameters defined in Soss model m, conditional on _data. Samples are drawn using Hamiltonial Monte Carlo (HMC) from the DynamicHMC.jl package.

This function is essentially a wrapper around DynamicHMC.mcmc_with_warmup(). Arguments reporter, ad_backend DynamicHMC docs here)

Arguments

rng: Random number generator.

m: Soss model.

_data: NamedTuple of data to condition on.

Keyword Arguments

N = 1000: Number of samples to draw.

method = logpdf: How to compute the log-density. Options are logpdf (delegates to logpdf of each component) or codegen (symbolic simplification and code generation).

ad_backend = Val(:ForwardDiff): Automatic differentiation backend.

reporter = DynamicHMC.NoProgressReport(): Specify logging during sampling. Default: do not log progress.

kwargs: Additional keyword arguments passed to core sampling function DynamicHMC.mcmc_with_warmup().

Returns an Array of Namedtuple of length N. Each entry in the array is a sample of parameters indexed by the parameter symbol.

Example

using Random
Random.seed!(42);
rng = MersenneTwister(42);

m = @model x begin
    β ~ Normal()
    yhat = β .* x
    y ~ For(eachindex(x)) do j
        Normal(yhat[j], 2.0)
    end
end

x = randn(50);
truth = rand(m(x=x));

post = dynamicHMC(rng, m(x=x), (y=truth.y,));
E_β = mean(getfield.(post, :β))

println("true β: " * string(round(truth.β, digits=2)))
println("Posterior mean β: " * string(round(E_β, digits=2)))

# output
true β: 0.3
Posterior mean β: 0.47

Soss.predictive — Method

predictive(m, xs...)

Returns a model transformed by adding xs... to arguments with a body containing only statements that depend on xs, or statements that are depended upon by children of xs through an open path. Unneeded arguments are trimmed.

Examples

m = @model (n, k) begin
    β ~ Gamma()
    α ~ Gamma()
    θ ~ Beta(α, β)
    x ~ Binomial(n, θ)
    z ~ Binomial(k, α / (α + β))
end;
predictive(m, :θ)

# output
@model (n, θ) begin
        x ~ Binomial(n, θ)
    end

Soss.prune — Method

prune(m, xs...)

Returns a model transformed by removing xs... and all variables that depend on xs.... Unneeded arguments are also removed.

Examples

m = @model n begin
    α ~ Gamma()
    β ~ Gamma()
    θ ~ Beta(α,β)
    x ~ Binomial(n, θ)
end;
prune(m, :θ)

# output
@model begin
        β ~ Gamma()
        α ~ Gamma()
    end

m = @model n begin
    α ~ Gamma()
    β ~ Gamma()
    θ ~ Beta(α,β)
    x ~ Binomial(n, θ)
end;
prune(m, :n)

# output
@model begin
        β ~ Gamma()
        α ~ Gamma()
        θ ~ Beta(α, β)
    end

Soss.PySymPyModule — Type

As type encoding a PyObject is unsafe without some hard works with reference counting, we simply use a Julia proxy to imitate the sympy module, e.g., sympy.attr = _pysympy.attr

Soss.after — Method

after(m::Model, xs...; strict=false)

Transforms m by moving xs to arguments. If strict=true, only descendants of xs are retained in the body. Otherwise, the remaining variables in the body are unmodified. Unused arguments are trimmed.

predictive(m::Model, xs...) = after(m, xs..., strict = true)

Do(m::Model, xs...) = after(m, xs..., strict = false)

Example

m = @model (n, k) begin
    β ~ Gamma()
    α ~ Gamma()
    θ ~ Beta(α, β)
    x ~ Binomial(n, θ)
    z ~ Binomial(k, α / (α + β))
end;
Soss.after(m, :α)

# output
@model (n, k, α) begin
        β ~ Gamma()
        θ ~ Beta(α, β)
        x ~ Binomial(n, θ)
        z ~ Binomial(k, α / (α + β))
    end

Soss.before — Method

before(m::Model, xs...; inclusive=true, strict=true)

Transforms m by retaining all ancestors of any of xs if strict=true; if strict=false, retains all variables that are not descendants of any xs. Note that adding more variables to xs cannot result in a larger model. If inclusive=true, xs is considered to be an ancestor of itself and is always included in the returned Model. Unneeded arguments are trimmed.

prune(m::Model, xs...) = before(m, xs..., inclusive = false, strict = false)

prior(m::Model, xs...) = before(m, xs..., inclusive = true, strict = true)

Examples

m = @model (n, k) begin
    β ~ Gamma()
    α ~ Gamma()
    θ ~ Beta(α, β)
    x ~ Binomial(n, θ)
    z ~ Binomial(k, α / (α + β))
end;
Soss.before(m, :θ, inclusive = true, strict = false)

# output
@model k begin
        β ~ Gamma()
        α ~ Gamma()
        θ ~ Beta(α, β)
        z ~ Binomial(k, α / (α + β))
    end

Soss.prior — Method

prior(m, xs...)

Returns the minimal model required to sample random variables xs.... Useful for extracting a prior distribution from a joint model m by designating xs... and the variables they depend on as the prior and hyperpriors.

Example

m = @model n begin
    α ~ Gamma()
    β ~ Gamma()
    θ ~ Beta(α,β)
    x ~ Binomial(n, θ)
end;
Soss.prior(m, :θ)

# output
@model begin
        β ~ Gamma()
        α ~ Gamma()
        θ ~ Beta(α, β)
    end